+++ packages/react/src/experimental/OneDataCollection/Settings/SettingsProvider.tsx
@@ -36,14 +36,13 @@ export interface DataCollectionSettingsContextType {
   settings: DataCollectionSettings
 }
 
-const DataCollectionSettingsContext =
-  createContext<DataCollectionSettingsContextType>({
-    setSettings: () => {},
-    settings: {
-      // To avoid circular dependency initializating the settings (the value is provided in the provider)
-      visualization: {} as VisualizationSettings,
-    },
-  })
+const DataCollectionSettingsContext = createContext({
+  setSettings: () => {},
+  settings: {
+    // To avoid circular dependency initializating the settings (the value is provided in the provider)
+    visualization: {} as VisualizationSettings,
+  },
+})
 
 export const useDataCollectionSettings = () => {
   const context = useContext(DataCollectionSettingsContext)
