+++ packages/react/src/lib/providers/datacollection/dataCollectionLocalStorageProvider.ts
@@ -1,16 +1,18 @@
-import { DataCollectionSettings } from "@/experimental/OneDataCollection/Settings/SettingsProvider"
-import { DataCollectionStorageProvider } from "./DataCollectionStorageProvider"
+import {
+  DataCollectionStorage,
+  DataCollectionStorageProvider,
+} from "./DataCollectionStorageProvider"
 
 export const dataCollectionLocalStorageProvider: DataCollectionStorageProvider =
   {
-    settings: {
-      get: async (keyName: string) => {
-        const key = `datacollection-${keyName}-settings`
-        return JSON.parse(localStorage.getItem(key) ?? "{}")
-      },
-      set: async (keyName: string, settings: DataCollectionSettings) => {
-        const key = `datacollection-${keyName}-settings`
-        localStorage.setItem(key, JSON.stringify(settings))
-      },
+    get: async (keyName: string) => {
+      const key = `datacollection-${keyName}-settings`
+      return JSON.parse(
+        localStorage.getItem(key) ?? "{}"
+      ) as DataCollectionStorage
+    },
+    set: async (keyName: string, settings: DataCollectionStorage) => {
+      const key = `datacollection-${keyName}-settings`
+      localStorage.setItem(key, JSON.stringify(settings))
     },
   }
